+++ src/app/page.tsx
@@ -13,11 +13,16 @@ export default function Home() {
   if (!session.data?.user) {
     return <div>Loading...</div>;
   }
+  console.log(session.data);
 
   return (
     <main className="flex min-h-screen flex-col items-center p-24">
       <h1 className="text-5xl bg-orangePrimary text-white">Home Page</h1>
-      <button className="text-white" onClick={() => signOut()}>
+      <p>{session.data.user.email}</p>
+      <button
+        className="border-orangePrimary border-2 m-2"
+        onClick={() => signOut()}
+      >
         Logout
       </button>
     </main>
